+++ src/main/java/com/example/lonua/user/model/entity/EmailVerify.java
@@ -1,4 +1,4 @@
-package com.example.lonua.user.model;
+package com.example.lonua.user.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
